Clean up resolvers and crud modules
- remove duplicated leftovers for tag crud functions
- move unboxed_items_collection resolver into box module
