+++ back/boxtribute_server/models/definitions/unboxed_items_collection.py
@@ -10,7 +10,6 @@ from .user import User
 
 
 class UnboxedItemsCollection(db.Model):
-    items_collection_type = "UnboxedItemsCollection"
     created_on = DateTimeField(null=True)
     created_by = UIntForeignKeyField(
         model=User,
